+++ variables.tf
@@ -35,12 +35,12 @@ variable "cluster_name" {
 }
 
 variable "cluster_version" {
-  description = "Kubernetes version to use for the Azure Kubernetes Service managed cluster, curently only \"1.21\" is supported."
+  description = "Kubernetes version to use for the Azure Kubernetes Service managed cluster, version \"1.21\" is supported & \"1.22\" is experimental."
   type        = string
 
   validation {
-    condition     = contains(["1.21"], var.cluster_version)
-    error_message = "Available versions are \"1.21\"."
+    condition     = contains(["1.22", "1.21"], var.cluster_version)
+    error_message = "Available versions are \"1.22\" & \"1.21\"."
   }
 }